Update database.yml to support dev environments
Currently, in `database.yml` most of the PostgreSQL-specific settings are only used for the `production` environment.  I'd suggest moving these up to the `default` section, and supplying sane Postgres defaults (hostname=localhost, port=5432).  

This is needed to, for example, run Lerna + PostgreSQL in a development environment on a VM or similar.

Also added a new Environment variable `RDS_DEVELOPMENT_DB_NAME` to allow for optionally customizing the name of the Postgres development DB in use for Lerna.