Basic example of job chaining
Built on top of the existing job.  So you could reasonably compose
different chains that used the same jobs in different sequences.

The main problem is that the two most costly jobs that benefit the most
from parallelization are solr and fedora creation, but those are the two
that we need to have logic about them both being complete.  So we either:
- sacrifice parallelism between solr and fedora creation to ensure logic
- allow the downstream job (e.g. AdminSet creation) to "spin" waiting for
  preconditions to arrive
- use a UX step/page with status polling that accomplishes the same thing

Spinning (check, sleep, check, sleep... eventually timeout) is
anti-pattern for queue mgmt.