Workaround child endpoint records not autosaving anymore.
These changes are only in the tests but it looks like this new behavior
shouldn't affect running applications since the jobs that create the
endpoints use #create_<name>_endpoint on an already saved account.

I could be wrong but it feels like the account and enpoint relationship
is reversed from what it should be and that this might be the source of the issue.