delete migration, rework User#add_default_group_memberships!
The MakeSureUsersHaveGroups migration was throwing errors when running on an existing Hyku application that already had data. Instead of making data changes in a migration (which is not best practice), replace it with a rake task that can be run when upgrading an existing app.

Also, rework User#add_default_group_memberships! to not:
1) run every time a user is saved
2) add users with the admin role to the admin group
Neither of those things are necessary. The new rake task takes care of updating existing admin users, and new tenants will always add admins to their admin group because of CreateAccount#add_initial_users