require either rights statement or rights notes be filled out on a generic work form
we do this by removing the intial requirement on the rights statement input. otherwise, [validateMetatdata](https://github.com/samvera/hyrax/blob/main/app/assets/javascripts/hyrax/save_work/save_work_control.es6\#L148-L155) remained false.

instead, on page load we verify if either field has data. if so, we carry on. if neither does, both are marked as required. from there, if one of the two fields is changed then it is marked as required while the other is removed as required.

if neither field has data and the form attempts to be saved (the save button will NOT be disabled), a popup will show up on the rights statement asking for it to be filled out.

TODO: make this unspecific to the generic work form.