♻️ Introduce Hyku::Application.theme_view_path_roots
With the introduction of [HykuKnapsack][1], we are adjusting how we
create instances of Hyku.  Namely we don't clone Hyku but instead we
incorporate Hyku as a submodule into a Knapsack.

The Knapsack is a Rails engine that is mounted in the Hyku application.
What this means is that when we want to

Prior to this commit, the only way to adjust themed views would have
been to add them to the Rails application (e.g. Hyku) directly.  Which
would work in a non-Knapsack ecosystem.

However, with Knapsack we need a means of saying "Hey, for themes we
want to be able to add/adjust views within the knapsack."  Hence this
change.

There should be a pairing commit in Knapsack that injects the
`HykuKnapsack::Engine.root` into the
`Hyku::Application.theme_view_path_roots` array.

[1]: https://github.com/samvera-labs/hyku_knapsack