Remove obsolete site roles and routes
This is a WIP.

The site/roles route is obsolete. This is an attempt to remove as much
of the obsolete logic as possible, as it has been replaced by the groups
with roles feature.

There is some confusion about which pieces may still be
in use, so further testing is needed to confirm that it doesn't remove
anything necessary.

There may be some unneeded remnants of prior roles logic in files such
as users.rb or ability.rb but this attempt cleans out some of the lower-hanging fruit.