🧹 Comment out check for reports
For this spec, if `Hyrax.config.analytics?` is enabled we need a legato
user to query the analytics.  Which creates a weird dependency that we
may not want in this integration test.

(This stuff is best tested in view tests; but here we are)

I opted to not remove the spec but instead favor the
`Hyrax.config.analytics?` which, if it is true likely means we have ENV
vars that are in play.