🧹 Update IIIF Print and AV gems
We were not seeing images loading correctly in the UV because we
introduced the `Hyku::WorksControllerBehavior` and IIIF Print did not
know about it.
  - https://github.com/scientist-softserv/iiif_print/commit/cad2cf33fc07ec2abb5b6834a858b5ed74903c12

We were seeing a respon to mismatch error in the Hyrax IIIF AV.
  - https://github.com/samvera-labs/hyrax-iiif_av/commit/4723b8fd0961fa179caf787532bd7922a44957ec

Also, the image that I tested was a phone image that apparently had
different ExifImageHeight/ExifImageWidth values than the
ImageHeight/ImageWidth values.  This was probably because modern phones
typically generate multiple images so the user can choose the best shot.
The exiftool_image_to_fits.xslt was proritizing the Exif values over the
regular values and would cause our image characterization to have weird
height and width values.  The change in this PR prioritizes the regular
over the Exif values.

Ref:
  - https://github.com/scientist-softserv/hykuup_knapsack/issues/68