🚧 Adding temporary hack so we can proceed
The problem we're facing is that the Valkyrie::IndexingAdapter registers
the Valkyrie::Indexing::Solr::IndexingAdapter as the default adapter,
however, this does not work in a multi-tenant environment.  The 'core'
is set to 'hyrax-valkyrie' and it needs to be the tenant's ID.  Also the
connection url needs the username and password in it.

See:
  - https://github.com/samvera/hyrax/blob/double_combo/lib/valkyrie/indexing/solr/indexing_adapter.rb#L71
  - https://github.com/samvera/hyrax/blob/double_combo/lib/valkyrie/indexing/solr/indexing_adapter.rb#L86

We put in the hack to get us past this issue.