🧹 Add redirects for work types
This commit will add redirects for the the work types so when viewing or
editing them, the user will be redirected to the Valkyrie (Resource)
version of the work type.  The won't work unless turbolinks is disabled.
We are opting for using an ENV variable to disable turbolinks for now so
it can be easily removed in the future.