🧹 Make the appropriate link generate
This commit will add a decorator to check the `human_readable_type` of
the given object is a Valkyrie migration object (which by convention
ends with 'resource') and adjust the generated link accordingly.  For
example, a GenericWorkResource should generate links like
`/concern/generic_work_resources/...` instead of
`/concern/generic_works/...`.

The models were updated to include the Hyrax::NestedWorks so the `Attach
Child` button would populate the appropriate work types.

Lastly, the `Gemfile.lock` was updated to include the latest version of
Hyrax `double_combo` branch.