🐛 Fix cropped banner images being overwritten
When cropper.js intercepts an updated banner image file and crops it,
the original filename gets changed to cropped.jpg which gets saved to
disk.  When another tenant crops their banner image, it also gets saved
as cropped.jpg and overwrites the previous tenant's banner image.  In
this commit, we prepend the tenant name so it would look something like
`shared-cropped.jpg` and `uindy-cropped.jpg` instead and avoids the
files from being overwritten.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/948
  - https://github.com/scientist-softserv/palni-palci/issues/746