🐛 Fix banner alt text require
The banner crop javascript was doing a `e.preventDefault();` which then
allowed the form being submitted without the alt text which is a
required field.  This commit will add a check to see if all the required
fields are filled in and if not then the function should not execute and
therefore not preventDefault.

Ref:
  - https://github.com/scientist-softserv/palni-palci/issues/948