Show PDF.js viewer when PDF.js FlipFlop is true
This is a workaround to unblock Mobius ingests. If the feature flipper to display PDF.js is true, we will display that viewer regardless of the property values of :show_pdf_viewer and :show_pdf_download_button.

We will revisit valkyrizing this feature fully in the future, when we start the iiif x valkyrie epic.

Issue:
- https://github.com/scientist-softserv/hykuup_knapsack/issues/185

Co-Authored-By: Kirk Wang <k3wang@gmail.com>