Fix duplicate thumbnails for PDFs
When text extraction fails on a PDF, the derivative job resubmits. This
was results in 5 Hyrax::FileMetadata objects for the same thumbnail
file, because jobs automatically resubmit 5 times.

In this PR, I commented out the failing extract_full_text method. I
also qualified the thumbnail creation on whether the thumbnail is already
there, in cases where we run more after creating the thumbnail.