🧹 Clean up featured collection UI
This commit will add a few more html tags to the featured collections,
both from the admin end and public facing end, to make it look more like
the featured works.  This will fix custom font issues where the titles
of the collections were not getting the correct font that was set.