🧹 Remove title from index and update translations
This commit will remove the title property from the index show page
because it already shows up in a different partial so it's redundant.
Also, there were a few missing translations so those have been updated.
It seemed that the translations were in two different spots so I
consolidated them so it's easier to manage.  I've updated the other
languages as well using the i18n Google translate task.