🧹 Remove override and add another for GA
This commit will remove the #process_files method override because we
have it contributed back to Hyrax.  It will also add an override to the
Hyrax::Statistic to get the correct ga_key to be used to find the
downloads stat from the Hyrax::Analytics::Result object for the
FileDownloadStat.