Override Sipity for Migrating Works
Sipity is unable to find the entities due to migrations. Errors manifest
in several situations, including:
- a fedora work searches for the entity by solr document. Due to the
lazy migration option, it can find the resource model rather than the
original model, resulting in an incorrect key proxy gid when trying
to find the entity.
- a migrated work may result in a new proxy gid when attempting to find
the sipity entity, resulting in the entity not being found. Sipity
entities need to be migrated as works are migrated.