Add custom query for IiifPrint
This query had previously been defined in Bulkrax. Bulkrax needed a change, and
the original query is still needed in IiifPrint.

Ideally this should be defined in IiifPrint directly, but the current setup
has the custom query for bulkrax being registered here. To avoid confusion,
this query is being added to Hyku since it is registered in Hyku.

If we determine a better way to register custom queries in gems, this can be moved
to IiifPrint.