File Set Indexing Bug fix
This is a partial fix for the file set PDF indexing bug. Indexing of the
resource relies on the file set PDF to be indexed if the file set exists.
The introduction of queued indexing caused errors in the indexing of the
work.

This fix removes the dependency issue, but only works for Valkyrie
resources. However, the methods being called would not have worked for
ActiveFedora resources anyway, so at this point we left the partial fix.

Future work should be done to ensure that this works for both
ActiveFedora and Valkyrie resources.