:bug: rescue google permission denied error
New tenants had analytics turned on by defaut when the global env variable was set. It should be set to false on a tenant level by default regardless if of the ENV var.

Additionally, when a tenant added configs that the site admin hasn't added as a service account to their Google Analytics, a Google::Cloud::PermissionDeniedError would break the site. This override will gracefully handle it by displaying an error on the dashboard.