🐛 A/V was not loading correctly
Hyrax changed the way derivatives are retrieved for audio and video
resources which was not compatible with the way Hyku generates IIIF
manifest urls.  This commit will simplify the code by using one av url
generator instead of separate ones for audio and video and pass the
mime_type in.

Ref
- https://github.com/notch8/hykuup_knapsack/issues/389
- https://github.com/samvera/hyrax/pull/6297