Set environment default for Hyku APP_NAME
**ISSUE**
When APP_NAME is not set in the environment (either directly before
launching the application, or via .env file), internal variables
such as `$HYKU_ADMIN_HOST` are set to invalid values,
causing admin routing to break.

I.E. when APP_NAME is not set, HYKU_ADMIN_HOST is set
to `admin-.localhost.direct`, which results in an illegal
hostname (because hostnames can not start or end with dashes).

**FIX**
Set the APP_NAME default in the .env file, which is loaded by
`docker compose` and interpolated into `docker-compose.yml`.
This ensures the samv value is used in both docker configuration
and internally in the Rails application.