Streamline Getting Started docs
The goal of this change is to provide a clear happy path for
developers to get a Hyku development environment running using
Docker.

This change intentionally moves or removes advanced options that
are not essential for setting up a basic sandbox environment.

Changes
* Remove references to production docker-compose setup in development section
* Remove obsolete instructions for fixing docker groups
* Update release tag for checkout to v6.2.0
* Emphasize Stack Car as happy path for DNS/TLS
* Moved all optional setup and config information to it's own section
* Added initial troubleshooting section
* [Temporary] Remove testing instructions since they fail

TODO: diagnose why local testing is not working, fix, and add
appropriate instructions.