🎁 hide_from_catalog_search attribute for flexible
This commit adds the hide_from_catalog_search attribute to the
Collection Resource class in flexible metadata. This adds the ability to
hide a collection from the catalog search results.

To hide a collection from the catalog search results, edit the
collection, go to the discovery tab on the form, check the box to
`Hide from catalog search`, save the changes.

Ref:
- https://github.com/notch8/adventist_knapsack/issues/892