🐛 Use post method for nested collection queries
This commit updates the Hyrax gem to the latest version of 5.0-flexible
that includes a bug fix in the Nested Collection Controller. Since it is
not based on catalog controller, so it does not inherit the POST
behavior we added there. This causes the parent_and_child_can_nest? to
do a GET request, which fails when too many ids are included in the
request. In this fix, we add an explicit setting of http_method prior
to doing this query.

Ref:
- https://github.com/notch8/adventist_knapsack/issues/1004