🐛 Add account cname to thumbnail img src
This commit will add the account cname to the thumbnail img tag's src
attribute.  In a search tenant we would only see /downloads/<id> which
doesn't work because it needs to know which tenant the document came
from.  Also fixed the cultural repository theme where the link was also
not displaying the full URL in a search tenant.

Ref:
- https://github.com/notch8/hykuup_knapsack/issues/480