🎁 Link to single sign-on documentation
This commit adds a link on the identity provider page to additional
information about single sign-on (SSO) located in the Samvera Confluence
user documentation guide.

Adds target="_blank" to all links on the page so that they open in a new
tab.

Adds `<code>` tags to the SSO attributes and the assertion consumer
service URL to improve readability.

Ref:
- https://github.com/notch8/hykuup_knapsack/issues/592