Move pages into a structure that can be merged into /docs
We want to move relevant pages into the /docs director of the main
Hyku repostiory. This makes it easier to version the documentation
with code releases.

This commit moves the source files into a structure that allows us
to merge them into the main repo without conflicts while retaining
the full wiki commit history.